Dynamically exclude FastAPI imports depending on pyodide
This commit makes it possible for the same shiny package to
work with both pyodide and FastAPI. It's done pretty crudely
here, the intent was to just do enough to let the pyodide branch
be merged back into main. After that, I'll revisit the way FastAPI
and pyodide plug into shiny.