Remove dependence on Starlette's lifespan mechanism
Reactive flush is what causes outputs and errors to get queued up
for sessions; a second step (unforunately also called "flush") is
necessary for those outputs and errors to be sent to clients. This
was achieved by subscribing to reactive._core.on_flushed with a
handler that triggers session._flush.

Shiny's reactive environment is global to a Python process. It is
possible for a single process to run multiple Shiny apps over time.
So the subscription to reactive._core.on_flushed was happening
during app startup and unsubscribed during app shutdown, as reported
via the Starlette lifespan feature.

However, if a Shiny app is mounted to a Starlette route, then its
lifespan never comes into play. This resulted in issue "#279:
reactivity failing when mounting app as route on starlette_app".

This commit fixes the problem by no longer using lifespan, but
instead, having each session subscribe/unsubscribe to on_flushed.