Stop using chunked encoding
When using --reload, the HTTP response for the homepage was using
Transfer-Encoding: chunked, which is more complicated for proxies
than it needs to be. This was happening because we had a script
autoinjection middleware that was modifying the body and then just
stripping the Content-Length header instead of making it the correct
value.