Using helper method approach to function and types for the user to make overloads manually
Before, one of the overloads was being made with input type `P`. This expands to `(*P.args, **P.kwargs) -> ...`.
Given our restrictions, `P.args` is equal to `*` and therefore is not in the function signature.
However, we need `*` to be there to allow for `_fn` to be passed in the other overloads.

In the case of `text` renderer, the function was being defined as a `Callable` whose args where `P`. This evaluated to no arguments. As a decorator, it needs to accept an argument of the function. This is incompatible.

However, with overloads, the overloaded functions can not really agree with each other as long as they agree with the base method. This allows us to "define" a dedorator who does not take an argument.

I find the function name to be too important to give up. Pursuing returning types and the renderer function for users to make overloads manually.