Seaborn plots should fill their output_plot
This fixes a bug introduced by #783, which was intended (in part)
to allow render.plot code to set plot sizes by using matplotlib
figure sizing. That code assumed that only explicit user code
would cause the figure size to differ from matplotlib's default
figure size (plt.rcParams["figure.figsize"]). But that's not true
for seaborn, which likes to set its own sizes, which we then
misidentified as being a high priority size directive that we
should obey.

https://seaborn.pydata.org/tutorial/function_overview.html#specifying-figure-sizes

This change removes the logic that looks at whether the figure
size changed. Instead, render.plot width/height can be set to 0
to cause the matplotlib figure size to be honored. However, we're
not going to be emphasizing that in our docs, at least for now.