Yield to give "synchronous" writes a chance to complete
During a synchronous phase of the event loop, we
can build up network messages that need to get to
the client. Because writes are naturally async,
we do some weird stuff to semi-execute them which
usually but doesn't always succeed. In the case
where the write doesn't succeed, they're held at
least until the next time the current task yields;
before this commit, that could be a long time
indeed, if the app is written synchronously (as
about 100% of Shiny apps are).

This commit yields to the event loop after
invalidating the session, which gives the writes
a chance to complete. This should take care of the
observed problem of progress messages not making
it (issue #1381). It also yields after each
successful @effect, in case insert_ui, custom
messages, etc. were sent during the @effect, and
also so outputs' automatic progress messages can
propagate to the client.