fix: Don't delete tempdir on garbage collection
If the `theme` is defined in the app UI, the theme object is gc'd
by the time the app is served (but after the UI has been rendered
to HTML tags). This causes the tempdir to be deleted, which then
cause the app to fail to serve. Since we need the tempdir to be
around for the life of the app, we should not delete it on gc.
We're now relying on the OS or Python to clean up tempdirs.