Revise accordion "open" logic for #2109
Code and docstring revised but not tested: Passing an empty iterable to ui.accordion now closes all panels.

A possible side-effect is that, if `restore_input` returns None, now the first panel will be open. (Old behavior was that all panels would be closed.) It's unclear whether `restore_input` would ever return None or what the expected behavior should be.